always use unsigned integers

+++ fuckshadows-csharp/Controller/Service/TCPRelay.cs
@@ -134,10 +134,10 @@ namespace Fuckshadows.Controller
         public const int ChunkOverheadSize = 16 * 2 /* two tags */ + AEADEncryptor.CHUNK_LEN_BYTES;
 
         // max chunk size
-        public const int MaxChunkSize = AEADEncryptor.CHUNK_LEN_MASK + AEADEncryptor.CHUNK_LEN_BYTES + 16 * 2;
+        public const uint MaxChunkSize = AEADEncryptor.CHUNK_LEN_MASK + AEADEncryptor.CHUNK_LEN_BYTES + 16 * 2;
 
         // In general, the ciphertext length, we should take overhead into account
-        public const int BufferSize = RecvSize + MaxChunkSize + 32 /* max salt len */ + 123 /* a random number */;
+        public const int BufferSize = RecvSize + (int)MaxChunkSize + 32 /* max salt len */ + 123 /* a random number */;
 
         public DateTime lastActivity;
